completed_rate

+++ spec/models/step_spec.rb
@@ -9,7 +9,7 @@ RSpec.describe Step, type: :model do
   describe "ActiveRecord validations" do
     it {should have_db_column(:name).of_type :string}
     it {should have_db_column(:content).of_type :text}
-    it {should have_db_column(:completed_rate).of_type :integer}
+    it {should have_db_column(:completed_rate).of_type :float}
     it {should have_db_column(:total_vote).of_type :integer}
     it {should have_db_column(:story_id).of_type :integer}
